Make snapshot_commit() takes in CommitRecord
