+++ storage/snapshot/snapshot.rs
@@ -12,7 +12,7 @@ use std::{
     ops::{Deref, DerefMut},
     sync::Arc,
 };
-
+use std::sync::atomic::Ordering;
 use bytes::byte_array::ByteArray;
 use error::typedb_error;
 use lending_iterator::LendingIterator;
@@ -37,6 +37,7 @@ use crate::{
     },
     MVCCStorage, StorageCommitError,
 };
+use crate::StorageCommitError::MVCCRead;
 
 macro_rules! get_mapped_method {
     ($method_name:ident, $get_func:ident) => {
@@ -212,6 +213,46 @@ where
     fn commit(self, commit_profile: &mut CommitProfile) -> Result<Option<SequenceNumber>, SnapshotError>;
 
     fn into_commit_record(self) -> CommitRecord;
+
+    fn set_initial_put_status(
+        &self,
+        storage: Arc<MVCCStorage<D>>,
+        storage_counters: StorageCounters,
+    ) -> Result<(), MVCCReadError>
+    {
+        for buffer in self.operations() {
+            let writes = buffer.writes();
+            let puts = writes.iter().filter_map(|(key, write)| match write {
+                Write::Put { value, reinsert, known_to_exist } => Some((key, value, reinsert, *known_to_exist)),
+                _ => None,
+            });
+            for (key, value, reinsert, known_to_exist) in puts {
+                let wrapped = StorageKeyReference::new_raw(buffer.keyspace_id, key);
+                if known_to_exist {
+                    debug_assert!(storage
+                        .get::<0>(
+                            self.iterator_pool(),
+                            wrapped,
+                            self.open_sequence_number(),
+                            storage_counters.clone()
+                        )
+                        .is_ok_and(|opt| opt.is_some()));
+                    reinsert.store(false, Ordering::Release);
+                } else {
+                    let existing_stored = storage
+                        .get::<BUFFER_VALUE_INLINE>(
+                            self.iterator_pool(),
+                            wrapped,
+                            self.open_sequence_number(),
+                            storage_counters.clone(),
+                        )?
+                        .is_some_and(|reference| &reference == value);
+                    reinsert.store(!existing_stored, Ordering::Release);
+                }
+            }
+        }
+        Ok(())
+    }
 }
 
 pub struct ReadSnapshot<D> {
@@ -390,7 +431,7 @@ impl<D> ReadableSnapshot for WriteSnapshot<D> {
             self.storage.iterate_range(self.iterator_pool(), range, self.open_sequence_number, storage_counters);
         SnapshotRangeIterator::new(storage_iterator, Some(buffered_iterator))
     }
-
+ 
     fn any_in_range<const PS: usize>(&self, range: &KeyRange<StorageKey<'_, PS>>, buffered_only: bool) -> bool {
         let buffered = self
             .operations
@@ -452,7 +493,14 @@ impl<D: DurabilityClient> CommittableSnapshot<D> for WriteSnapshot<D> {
         if self.operations.is_writes_empty() && self.operations.locks_empty() {
             Ok(None)
         } else {
-            match self.storage.clone().snapshot_commit(self, commit_profile) {
+            self.set_initial_put_status(self.storage.clone(), commit_profile.storage_counters())
+                .map_err(|error| SnapshotError::Commit {
+                    typedb_source: MVCCRead { name: self.storage.name.clone(), source: error } 
+                })?;
+            commit_profile.snapshot_put_statuses_checked();
+            let storage = self.storage.clone();
+            let commit_record = self.into_commit_record();
+            match storage.clone().commit(commit_record, commit_profile) {
                 Ok(sequence_number) => Ok(Some(sequence_number)),
                 Err(error) => Err(SnapshotError::Commit { typedb_source: error }),
             }
@@ -611,7 +659,14 @@ impl<D: DurabilityClient> CommittableSnapshot<D> for SchemaSnapshot<D> {
         if self.operations.is_writes_empty() && self.operations.locks_empty() {
             Ok(None)
         } else {
-            match self.storage.clone().snapshot_commit(self, commit_profile) {
+            self.set_initial_put_status(self.storage.clone(), commit_profile.storage_counters())
+                .map_err(|error| SnapshotError::Commit {
+                    typedb_source: MVCCRead { name: self.storage.name.clone(), source: error }
+                })?;
+            commit_profile.snapshot_put_statuses_checked();
+            let storage = self.storage.clone();
+            let commit_record = self.into_commit_record();
+            match storage.commit(commit_record, commit_profile) {
                 Ok(sequence_number) => Ok(Some(sequence_number)),
                 Err(error) => Err(SnapshotError::Commit { typedb_source: error }),
             }
